move searchable to meta tab in field detail
